style: change following code review
